change directory structure of panorama configs
