+++ src/components/PannellumPanorama.vue
@@ -5,7 +5,7 @@
 <script>
   import 'pannellum'
   import 'pannellum/build/pannellum.css'
-  import axios from 'axios'
+  import configs from '../config/PanoramaConfig.js'
   
   export default {
     mounted() {
@@ -13,78 +13,6 @@
     },
     methods: {
       initPanorama() {
-        const configs = [{
-          id: 0,
-          autoLoad: true,
-          panorama: 'http://127.0.0.1:8080/static/panos/pano_0.jpg',
-          hotSpots: [
-            {
-              pitch:  -3.562,
-              yaw: -14.748,
-              type: 'scene',
-              clickHandlerFunc: this.getNextPanorama,
-              clickHandlerArgs: ['+']
-            }
-          ]
-        },
-          {
-            id: 1,
-            autoLoad: true,
-            panorama: 'http://127.0.0.1:8080/static/panos/pano_1.jpg',
-            hotSpots: [
-              {
-                pitch:  -25.727,
-                yaw: -95.330,
-                type: 'scene',
-                clickHandlerFunc: this.getNextPanorama,
-                clickHandlerArgs: ['+']
-              },
-              {
-                pitch:  -16.714,
-                yaw: 100.570,
-                type: 'scene',
-                clickHandlerFunc: this.getNextPanorama,
-                clickHandlerArgs: ['-']
-              }
-            ]
-          },
-          {
-            id: 2,
-            autoLoad: true,
-            panorama: 'http://127.0.0.1:8080/static/panos/pano_2.jpg',
-            hotSpots: [
-              {
-                pitch:  -14.975,
-                yaw: -7.043,
-                type: 'scene',
-                clickHandlerFunc: this.getNextPanorama,
-                clickHandlerArgs: ['+']
-              },
-              {
-                pitch:  -11.200,
-                yaw: -149.777,
-                type: 'scene',
-                clickHandlerFunc: this.getNextPanorama,
-                clickHandlerArgs: ['-']
-              }
-            ]
-          },
-          {
-            id: 3,
-            autoLoad: true,
-            panorama: 'http://127.0.0.1:8080/static/panos/pano_3.jpg',
-            hotSpots: [
-              {
-                pitch:  -16.203,
-                yaw: 179.270,
-                type: 'scene',
-                clickHandlerFunc: this.getNextPanorama,
-                clickHandlerArgs: ['-']
-              }
-            ]
-          }
-          ]
-
         this.viewer = pannellum.viewer(this.$refs.panorama, {
           scenes: {},
           "sceneFadeDuration": 1000,
@@ -100,33 +28,7 @@
         this.viewer.on('mousedown', (event) =>{
           console.log(this.viewer.mouseEventToCoords(event));
         });
-
       },
-
-      getNextPanorama(event, args) {
-        const current_pano = this.viewer.getConfig().panorama
-        const direction = args[0]
-
-        // 发送HTTP请求给后端: 传当前场景、参数给后端
-        axios.get('http://localhost:8080/api/next-pano', {
-          params: {
-            current_pano: current_pano,
-            direction: direction
-          },
-        })
-        .then(response => {
-          console.log(response.data);
-
-          // 切换到下张全景图
-          const next_pano_id = response.data["next_pano_id"]
-
-          // 加载新场景
-          this.viewer.loadScene(next_pano_id, this.viewer.getPitch(), this.viewer.getYaw(), this.viewer.getHfov())
-        })
-        .catch(error => {
-          console.error('请求失败', error);
-        });
-      }
     }
   };
 </script>
